+++ src/main/java/ru/yourteam/filmorate/controller/DirectorFilmsController.java
@@ -1,8 +1,8 @@
+// Контроллер для работы с фильмами и их режиссёрами.
 package ru.yourteam.filmorate.controller;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Min;
-import jakarta.validation.constraints.NotBlank;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -17,9 +17,6 @@ import java.util.Collection;
 import ru.yourteam.filmorate.dto.FilmDto;
 import java.util.List;
 
-/**
- * Контроллер для работы с фильмами и их режиссёрами.
- */
 @Validated
 @RestController
 @RequestMapping("/directors")
@@ -56,7 +53,7 @@ public class DirectorFilmsController {
 
     // изменение режиссера
     @PutMapping
-    public ResponseEntity<DirectorDto> update(@Valid @RequestBody DirectorDto director) {
+    public ResponseEntity<DirectorDto> update(@RequestBody DirectorDto director) {
         return ResponseEntity.ok(directorServiceImpl.update(director));
     }
 
@@ -70,7 +67,7 @@ public class DirectorFilmsController {
     @GetMapping("/{id}/films")
     public List<FilmDto> getFilms(
         @PathVariable long id,
-        @RequestParam("sort") @NotBlank String sort) {
+        @RequestParam("sort") String sort) {
         return directorServiceImpl.getFilmsByDirector(id, SortMode.from(sort)); // SortMode.parse/validate бросит 400
     }
 }