+++ src/main/java/ru/yourteam/filmorate/service/SearchService.java
@@ -4,9 +4,9 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
-import ru.yourteam.filmorate.dto.search.FilmSearchDto;
-import ru.yourteam.filmorate.mapper.search.FilmSearchMapper;
-import ru.yourteam.filmorate.repository.SearchRepository;
+import ru.yourteam.filmorate.dal.repositories.searchRepository.SearchRepository;
+import ru.yourteam.filmorate.dto.searchDto.FilmSearchDto;
+import ru.yourteam.filmorate.mappers.searchMapper.FilmSearchMapper;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -23,19 +23,16 @@ public class SearchService {
         if (query != null) {
             switch (by.toLowerCase()) {
                 case "director" -> {
-                    return searchRepository.getAllSortedByRatingFilms(Type.DIRECTOR, query).stream()
-                        .map(FilmSearchMapper::mapToFilmSearchDto)
-                        .collect(Collectors.toList());
+                    return searchRepository.getAllSortedByRatingFilms(Type.DIRECTOR, query).stream().map(FilmSearchMapper::mapToFilmSearchDto).
+                        collect(Collectors.toList());
                 }
                 case "title" -> {
-                    return searchRepository.getAllSortedByRatingFilms(Type.TITLE, query).stream()
-                        .map(FilmSearchMapper::mapToFilmSearchDto)
-                        .collect(Collectors.toList());
+                    return searchRepository.getAllSortedByRatingFilms(Type.TITLE, query).stream().map(FilmSearchMapper::mapToFilmSearchDto).
+                        collect(Collectors.toList());
                 }
                 case "director,title" -> {
-                    return searchRepository.getAllSortedByRatingFilms(Type.ALL, query).stream()
-                        .map(FilmSearchMapper::mapToFilmSearchDto)
-                        .collect(Collectors.toList());
+                    return searchRepository.getAllSortedByRatingFilms(Type.ALL, query).stream().map(FilmSearchMapper::mapToFilmSearchDto).
+                        collect(Collectors.toList());
                 }
                 default -> throw new ResponseStatusException(
                     HttpStatus.BAD_REQUEST,
@@ -43,9 +40,8 @@ public class SearchService {
                 );
             }
         }
-        return searchRepository.getAllSortedByRatingFilms(Type.NOTHING, null).stream()
-            .map(FilmSearchMapper::mapToFilmSearchDto)
-            .collect(Collectors.toList());
+        return searchRepository.getAllSortedByRatingFilms(Type.NOTHING, null).stream().map(FilmSearchMapper::mapToFilmSearchDto).
+            collect(Collectors.toList());
 
     }